app/requirements.txt
absl-py==2.3.1
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
datasets==4.0.0
dill==0.3.8
distro==1.9.0
Django==4.2.3
djangorestframework==3.16.1
evaluate==0.4.5
filelock==3.19.1
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.3.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
keras==3.11.3
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
mysqlclient==2.2.7
namex==0.1.0
networkx==3.5
numpy==2.1.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai==1.101.0
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.8.29
requests==2.32.5
rich==14.1.0
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.1
termcolor==3.1.0
tf_keras==2.19.0
threadpoolctl==3.6.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.0
triton==3.4.0
trustpilot-scraper==0.10
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
